requirements.txt
Flask
flask-cors
python-dotenv
Werkzeug
groq
gunicorn
